Reorganização dos arquivos | Acrescimo de uma ViewModel e Factory | Uptade de icon n c
