+++ app/src/main/java/com/ifam/pdm/starbemapp/model/AtividadeDao.java
@@ -5,7 +5,7 @@ import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 
-import com.ifam.pdm.starbemapp.BDAtividades;
+import com.ifam.pdm.starbemapp.model.BDAtividades;
 
 import java.util.ArrayList;
 import java.util.List;
