+++ app/src/main/java/com/ifam/pdm/starbemapp/viewmodel/HomeViewModel.java
@@ -1,0 +1,4 @@
+package com.ifam.pdm.starbemapp.viewmodel;
+
+public class HomeViewModel {
+}
